pre-processor: no-bg-clamp + optional edge noise operation
A --no-bg-clamp option and "no_bg_clamp" property let us control
whether the pre-processor will explicitly set all background pixels to
have a depth according to `background_depth_m`. This is useful when
creating synthetic test recordings where we want background objects
and a floor.

The --no-flip option is now possible to control via a pre-processor
config with a "no_flip" property.

The noise that we add to the silhouette of the body is now an optional
noise operation called "foreground-edge-swizzle" and the pre-processor
configs will be updated to include this as the first operation to
maintain the same behaviour.