joint_inferrer: optimize joints_inferrer_infer_fast()
This makes a few optimizations:
1) Lots of intermediate storage is now std::vector based and the vectors
   hang off the struct joints_inferrer to be re-used.
2) With a new sparse representation of spans this introduces a new
   span merging algorithm, using a 'runs' array similar to
   cluster_codebook_classified_points() to tracking temporary links
   between IDs. In particular this iterates through lines and only
   ever looks to merge with spans that are one line above the current
   span - whereas before for N spans we did N x N merge comparisons.
3) Calculating the nearest point to the average center of a cluster is
   done based on the squared distance; avoiding lots of unnecessary
   square root calculations.

To avoid changing the interface at this point Joint still has an
embedded array of linked lists but the code for building the
InferredJoints result is self contained at the end of the function
and will be easy to replace. One minor change to Joint was to just
change the Joint** joints; member into a Joint* joints[]; array so
the allocation can be combined into the allocation for the
InferredJoints struct.