train_rdt: require a --log* option
There is now no default logging to stderr to encourage that a persistent
log.

file be written when training.  This is considering that training may
run for days, and so a failure could take a long time to reproduce but
that may be unnecessary if a log indicates the failure.

Logging to stderr can be enabled explicitly via --log-stderr.