context: Delay codebook updates on the bounding cube of untracked people
Previously, we would delay *all* codebook updates for a period after a
person that was previously tracked wasn't tracked. Now we delay updates
only on the bounding cube of the untracked person. This should better
allow us to track multiple people in the future (as well as making the
code flow a bit easier to understand).