install_glimpse: on osx only install plugin as a .bundle
Since looking at updating what version of Meson we build with it was
noticed that our OSX builds of the Unity plugin where actually only
working because of the absolute rpath that was added to the
glimpse-unity-plugin so that the various .dylib deps could be resolved
within the Meson build directory. Since Meson now adds relative rpaths
(using @loader_path) then the fact that we rename all the .dylib files
means our plugin dependencies aren't resolved at link time.

Since Unity only considers .bundle files as plugins we now only rename
libglimpse-unity-plugin.dylib to glimpse-unity-plugin.bundle and we
also use install_name_tool on OSX to change the id of
glimpse-unity-plugin.bundle according to its new name and we give all
shared libraries and the plugin an rpath of @loader_path, consistent
with adding $ORIGIN on Linux, so dependencies are resolved adjacent to
the location of the plugin.