viewer: remove features related to target sequences
We originally created the target sequences api because we thought it
would solve an application requirement we had for matching tracked user
poses against reference poses.

Since developing that application further we've found that target
sequences as supported by the gm_target_ api don't really solve anything
that the application couldn't easily handle itself and in fact that
would be more practical since there's application-specific meta data
that could then naturally be part of the target description itself.

For now this just removes the viewer UI relating to target sequences
since the application is currently still dependant on the gm_target
api but it's convenient to remove the viewer UI while it depends on API
we want to remove.