Remove older _get_skeleton / _get_prediction apis
This removes the older apis for getting tracking skeletons or
predictions without specifying a person_id - updating or removing uses
of the older api.

Various glimpse_viewer features (that we haven't been using recently anyway)
that used gm_tracking_get_skeleton have been removed.

Notably one of the users of gm_tracking_get_skeleton was the
make_target_sequences tool but since we're planning to remove the
gm_targets api it didn't make sense to make corresponding changes to the
gm_targets implementation. This renames the tool to 'mocap' and it now
supports outputting tracked poses for multiple people. The tool is now
considered unrelated from gm_targets.

So that it's possible to query the tracked skeletons that are directly
related to a gm_tracking object we have these new apis:

  gm_tracking_get_tracked_people_ids()
  gm_tracking_get_tracked_skeleton_for_person()

Since it seemed like a slightly misleading name and to maintain
consistency gm_context_get_people() was renamed to
gm_context_get_people_ids(). This api now also doesn't return an
malloc'd array, it expects to be passed in a sized array and the
function returns how many ids were written to the array (this is a
little easier to bind into Unity).