Create re-useable 'shell' for graphical tools
This creates a callback based abstraction over glfw and glfm to make it
more practical to create multiple graphical tools that work across
operating systems.

This adds a bare-bones glimpse_shell_test executable that just creates a
blank window to test the abstraction.

glimpse_viewer has been updated to use this and tested on OSX, iOS and
Android.