Add Metal API support to shell
The aim is to be able to use Metal compute shaders for evaluating our
decision trees.

Although this does make various corresponding changes to glimpse_viewer
including reorganising to group all the GL-specific code; this does not
add any support for visualising the debug cloud / geometry via Metal
(the debug visualisations will be skipped if running with Metal).

This includes an update to imgui including a Metal backend.