Revert "annotate_bone_map: Support multiple heads/tails and bone names"
This reverts commit d3d9c3cac0d9261522401ad1be82f4dfa7e99259.

Conceptually defining that bones have multiple 'heads' and 'tails' is
confusing (at least to me) and seems to complicate context code which
doesn't seem worthwhile when we aren't making use of this feature
currently.

One idea is that if we want to generalise retargetting/mapping one
skeleton topology to another then we can add a mapping feature which can
support deriving joints as an interpolation of other joints, but this
can be an orthogonal feature so our skeleton representation can remain
simple.