context: revert back to a simpler gm_bone with one head/tail
This effectively reverts 0273f9fdc, but done manually since various
changes have been made since then.

On the one hand the terminology was odd (to me) since it didn't seem to
make sense to say a bone has multiple 'heads' and 'tails', but also the
representation made reasoning about skeleton manipulations more complex
than I think it was before (at least just counting lines of code it
seems like it was less complex before).

E.g. for me I felt this was getting in the way of looking at utilizing
skeletal information to make a second pass over labels to mask pixels
according to physical body-topology constraints.

Incidentally from reviewing this I became pretty convinced that the
logic for sanitize_bone_rotations() probably doesn't do what we want
due to the way that we compare the angles of corresponding bones (between
separate skeletons) in world coordinates.

Incidentally I also noticed that predict_skeleton_for_history was
modifying the bone valid state for the closest matching skeleton in the
history, where it was intending to modify the valid state of the
skeleton being predicted (fixed).

Although for now this means we can't load arbitrary bone-map topologies
(notably unused currently) I think we could re-introduce something
similar that's implemented in terms of a mapping operation that takes a
skeleton with one topology and spits out another one in a new topology.
This could be a more orthogonal feature that doesn't require changing
how skeletons and bones are generally represented (except allowing
multiple sets of bone_info). In this case it would just be the mapping
descriptor itself that would describe deriving joints as a blend of
other joints as necessary.