Split up joint inference state
This splits joints_inferrer into joints_inferrer and
joints_inferrer_state so that the joint_map can be shared from one place
(the joints_inferrer) but all other intermediate state is moved to
joints_inferrer_state which is now associated with gm_tracking_impl.

The intent is to enable some debug visualisations of the joint
inferrence work (such as the clustering) so we want to be able to
preserve some of the intermediate work under a tracking object but not
duplicate the joint_map.